arreglando BaseDao para que room compile :v
arreglando tipado de RutinaDetalleViewmodel... fijarse en el modelo de datos de ejercicio

la screen esta rota

+++ app/src/main/java/cl/duocuc/gymcel/presentacion/ui/screens/RutinaDetalleScreen.kt
@@ -83,7 +83,7 @@ data class EjercicioApiTemp(
 @Composable
 fun RutinaDetalleScreen(
     navController: NavController,
-    rutinaId: Int?,
+    rutinaId: Long?,
     viewModel: RutinaDetalleViewModel = viewModel()
 ) {
     val rutina by viewModel.rutina.collectAsState()
